requirements.txt
dlib==19.24.1
face_recognition==1.3.0
face_recognition_models==0.3.0
opencv-python==4.13.0.92
numpy==1.26.4
Pillow==12.2.0
